master_langchain/requirements.txt
langchain
langchain-cohere
cohere
pypdf
chromadb
streamlit
python-dotenv
faiss-cpu
